+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.4.1] 2019-09-06
 ### Fixed:
 - minor typo
+- Handle missing default profile gracefully (#57)
 
 ## [0.4.0] 2019-04-14
 ### Added: